refactor(package/rolldown): resolve circular dependency
